+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/berry-burst/disguise-berry-burst-details.vue
@@ -67,7 +67,7 @@
 
 <script lang="ts">
 import { berryImage, mainskillImage } from '@/services/utils/image-utils'
-import { skillLevelBadgeText } from '@/services/utils/skill-level-utils'
+import { skillLevelBadgeText } from '@/services/utils/skill-display-utils'
 import { useTeamStore } from '@/stores/team/team-store'
 import type { MemberProductionExt } from '@/types/member/instanced'
 import { BerryBurstDisguise, MathUtils, compactNumber } from 'sleepapi-common'