+++ frontend/src/components/calculator/results/member-results/member-production-header/member-production-skill-details/cooking-power-up-s/minus-cooking-power-up-s-details.vue
@@ -68,7 +68,7 @@
 <script lang="ts">
 import { mainskillImage } from '@/services/utils/image-utils'
 import { usePokemonStore } from '@/stores/pokemon/pokemon-store'
-import { skillLevelBadgeText } from '@/services/utils/skill-level-utils'
+import { skillLevelBadgeText } from '@/services/utils/skill-display-utils'
 import { useTeamStore } from '@/stores/team/team-store'
 import type { MemberProductionExt } from '@/types/member/instanced'
 import { CookingPowerUpSMinus, MathUtils, compactNumber, isPlusOrMinus } from 'sleepapi-common'